fix: "수익률 단위 변경"
double 상태로 문자열로 바뀌던 기능을 long 으로 형변환하여 소수점을 없앰

+++ src/main/java/lotto/domain/CompareLotto.java
@@ -69,6 +69,6 @@ public class CompareLotto {
 
         double totalYield = (receiveMoney / originalMoney) * 100;
 
-        return "총 수익률은 " + totalYield + "%입니다.";
+        return "총 수익률은 " + (long)totalYield + "%입니다.";
     }
 }
